+++ ormar/models/mixins/__init__.py
@@ -8,6 +8,7 @@ from ormar.models.mixins.alias_mixin import AliasMixin
 from ormar.models.mixins.excludable_mixin import ExcludableMixin
 from ormar.models.mixins.merge_mixin import MergeModelMixin
 from ormar.models.mixins.prefetch_mixin import PrefetchQueryMixin
+from ormar.models.mixins.pydantic_mixin import PydanticMixin
 from ormar.models.mixins.save_mixin import SavePrepareMixin
 
 __all__ = [
@@ -16,4 +17,5 @@ __all__ = [
     "PrefetchQueryMixin",
     "SavePrepareMixin",
     "ExcludableMixin",
+    "PydanticMixin",
 ]